+++ Appraisals
@@ -1,31 +1,19 @@
-appraise "spree-3-2" do
-  gem "spree", "~> 3.2.0"
-  gem "spree_auth_devise", "~> 3.2.0"
-end
-
-appraise "spree-3-3" do
-  gem "spree", "~> 3.3.0"
-  gem "spree_auth_devise", "~> 3.3.0"
-end
-
-appraise "spree-3-4" do
-  gem "spree", '~> 3.4.0'
-  gem 'spree_auth_devise', github: 'spree/spree_auth_devise', branch: 'master'
-end
-
-appraise 'spree-3-5' do
-  gem 'spree', '~> 3.5.0'
-  gem 'spree_auth_devise', '~> 3.3.0'
-end
-
-appraise 'spree-3-6' do
-  gem 'spree', '~> 3.6.0'
-  gem 'spree_auth_devise', '~> 3.3.0'
-end
-
 appraise 'spree-3-7' do
   gem 'spree', '~> 3.7.0'
+  gem 'spree_backend', '~> 3.7.0'
   gem 'spree_auth_devise', '~> 3.5.0'
+end
+
+appraise 'spree-4-0' do
+  gem 'spree', '~> 4.0.0'
+  gem 'spree_backend', '~> 4.0.0'
+  gem 'spree_auth_devise'
+end
+
+appraise 'spree-4-1' do
+  gem 'spree', '~> 4.1.0'
+  gem 'spree_backend', '~> 4.1.0'
+  gem 'spree_auth_devise'
 end
 
 appraise 'spree-master' do
