--- app/models/spree/user_decorator.rb
@@ -1,5 +1,0 @@
-module Spree
-  Spree.user_class.class_eval do
-    alias_attribute :roles, :spree_roles
-  end
-end
