+++ app/models/spree_admin_roles_and_access/spree/user_decorator.rb
@@ -1,0 +1,11 @@
+module SpreeAdminRolesAndAccess
+  module Spree
+    module UserDecorator
+      def self.prepended(base)
+        base.alias_attribute :roles, :spree_roles
+      end
+    end
+  end
+end
+
+::Spree::User.prepend SpreeAdminRolesAndAccess::Spree::UserDecorator
